application/requirements.txt
flask==3.0.3
requests==2.31.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-flask==0.45b0
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation-logging==0.45b0